refactor: 분리한 Divider 통합
